src/requirements-deepface.txt
absl-py==2.1.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
deepface==0.0.93
docker-pycreds==0.4.0
easydict==1.13
eval_type_backport==0.2.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
flit_core @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/flit_core/flit_core-3.9.0
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.68.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
imageio==2.36.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/packaging/packaging-24.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
protobuf==5.29.0
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/setuptools/setuptools-70.0.0
setuptools-scm @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/setuptools_scm/setuptools_scm-8.1.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stringzilla==3.10.10
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/tomli/tomli-2.0.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.1.3
wheel @ file:///dev/shm/Python/3.12.3/GCCcore-13.3.0/wheel/wheel-0.43.0
wrapt==1.17.0
